New CR handling
Much simpler handling of CR: have internal KSW handler "sideline" CRs it
receives, so we can arrange to "receive" it at a more convenient time
(after setting the cursor to the end of the line, and any other
cleanup).

Eliminates EOL position saving, and stack pointer saving to handle
ProDOS's set-X-and-backspace "prompt reset" thing. Also simplifies
checking for that backspace thing, since we know for a fact it
should be a CR, so if it's a BS then we don't have to additionally
check if X changed.